+++ test/req/request_test.exs
@@ -1,6 +1,6 @@
 defmodule Req.RequestTest do
   use ExUnit.Case, async: true
-  doctest Req.Request, only: [register_option: 2]
+  doctest Req.Request, only: [register_options: 2]
 
   setup do
     bypass = Bypass.open()