+++ CHANGELOG.md
@@ -207,7 +207,7 @@ See "Adapter" seciton in `Req.Request` module documentation for more information
 
   * Add `Req.Request.adapter` field
 
-  * Add `Req.Request.register_option/2`
+  * Add `Req.Request.register_options/2`
 
   * Rename `put_if_modified_since` step to `cache`